dom/tests/mochitest/ajax/offline/test_redirectUpdateItem.html

<html xmlns="http://www.w3.org/1999/xhtml" manifest="http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/redirects.cacheManifest">
<head>
<title>Entries redirection handling during update test</title>

<script type="text/javascript" src="/MochiKit/packed.js"></script>
<script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript" src="/tests/dom/tests/mochitest/ajax/offline/offlineTests.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script class="testbody" type="text/javascript">

/**
 */

// The manifest is always the same, we just update its version comment
var manifestVersion2Content =
  "CACHE MANIFEST\n" +
  "# v2\n" +
  "\n" +
  "http://localhost:8888/tests/SimpleTest/SimpleTest.js\n" +
  "http://localhost:8888/MochiKit/packed.js\n" +
  "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/offlineTests.js\n" +
  "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/explicitRedirect.sjs";

var manifestVersion3Content =
  "CACHE MANIFEST\n" +
  "# v3\n" +
  "\n" +
  "http://localhost:8888/tests/SimpleTest/SimpleTest.js\n" +
  "http://localhost:8888/MochiKit/packed.js\n" +
  "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/offlineTests.js\n" +
  "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/explicitRedirect.sjs";


var gCurrentManifestVersion = 1;

function manifestCached()
{
  OfflineTest.checkCache(
    "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/dynamicRedirect.sjs", false);
  OfflineTest.checkCache(
    "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/explicitRedirect.sjs", true);
  OfflineTest.is(gCurrentManifestVersion, 1, "Cached event for manifest version one");

  // Now add one dynamic entry (now with content overriden redirect sjs)
  applicationCache.mozAdd(
    "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/dynamicRedirect.sjs");

  // Wait for the dynamic entry be added to the cache...
  OfflineTest.waitForAdd(
    "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/dynamicRedirect.sjs",
    function() {
      // ...check it is there...
      OfflineTest.checkCache(
        "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/dynamicRedirect.sjs", true);

      // ...delete content of the dynamic entry from the server, now we get the redirect...
      OfflineTest.deleteData(
        "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/dynamicRedirect.sjs");

      // ...update manifest to the new version on the server...
      OfflineTest.putData(
        "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/redirects.cacheManifest",
        "text/cache-manifest", manifestVersion2Content);
      gCurrentManifestVersion = 2;

      // ...and finally invoke the cache update.
      applicationCache.update();
    });
}

function manifestUpdated()
{
  switch (gCurrentManifestVersion)
  {
    case 2:
      // Check the dynamic entry was removed from the cache (because of the redirect)...
      OfflineTest.checkCache(
        "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/dynamicRedirect.sjs", false);

      // ...return back redirect for the explicit entry...
      OfflineTest.deleteData(
        "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/explicitRedirect.sjs");

      // ...update the manifest to the third version...
      OfflineTest.putData(
        "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/redirects.cacheManifest",
        "text/cache-manifest", manifestVersion3Content);
      gCurrentManifestVersion = 3;

      // ...and invoke the cache update, now we must get error.
      applicationCache.update();
      break;

    case 3:
      OfflineTest.ok(false, "Update failed for third version of the manifest");
      finish();
      break;
  }
}

function manifestError()
{
  switch (gCurrentManifestVersion)
  {
    case 1:
      OfflineTest.ok(false, "Error not expected when caching the first version of the manifest");
      finish();
      break;
    case 2:
      OfflineTest.ok(false, "Error not expected when updating to second version of the manifest");
      finish();
      break;
    case 3:
      OfflineTest.ok(true, "Error expected when updating to third version of the manifest");
      finish();
      break;
  }
}

function finish()
{
  OfflineTest.teardown();
  OfflineTest.finish();
}

SimpleTest.waitForExplicitFinish();

if (OfflineTest.setup()) {
  applicationCache.onerror = OfflineTest.priv(manifestError);
  applicationCache.onupdateready = OfflineTest.priv(manifestUpdated);
  applicationCache.oncached = OfflineTest.priv(manifestCached);

  // Override sjs redirects on the server, it will now return 200 OK and the content
  OfflineTest.putData(
      "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/explicitRedirect.sjs",
      "text/html",
      "<html><body>Explicit page</body></html>");
  OfflineTest.putData(
      "http://localhost:8888/tests/dom/tests/mochitest/ajax/offline/dynamicRedirect.sjs",
      "text/html",
      "<html><body>Dynamic page</body></html>");
}

</script>

</head>

<body>
</body>
</html>
